Corrections in package and class name
