+++ src/main/scala/br/usp/ime/fllsouto/diningPhilosophers/chandyMisra/ChandyMisraDinnerMaster.scala
@@ -1,6 +1,6 @@
-package br.usp.ime.fllsouto.dinningActors.resourceHierarchy
+package br.usp.ime.fllsouto.diningPhilosophers.chandyMisra
 
-import br.usp.ime.fllsouto.dinningActors._
+import br.usp.ime.fllsouto.diningPhilosophers._
 import akka.actor._
 import scala.concurrent.duration._
 import akka.pattern.gracefulStop
@@ -10,7 +10,7 @@ import scala.concurrent.{Await, ExecutionContext, Future}
 case class TerminateDinner(philos: List[ActorRef], forks: IndexedSeq[ActorRef], logger: ActorRef)
 
 
-class ResourceHierarchyDinnerMaster(duration: FiniteDuration, fakeRandom: Boolean) extends Actor {
+class ChandyMisraDinnerMaster(duration: FiniteDuration, fakeRandom: Boolean) extends Actor {
   import context._
   val r = scala.util.Random
 
@@ -28,17 +28,19 @@ class ResourceHierarchyDinnerMaster(duration: FiniteDuration, fakeRandom: Boolea
 
       self ! PoisonPill
 
-      case undefMessage => println("[%s][%s] Actor: %s, State: %s, Message: %s".format("ResourceHierarchy", "ResourceHierarchyDinnerMaster", self.path.name, "receive", undefMessage))
+    case undefMessage => println("[%s][%s] Actor: %s, State: %s, Message: %s".format("ChandyMisra", "ChandyMisraDinnerMaster", self.path.name, "receive", undefMessage))
   }
 
   private def run() {
-    val fn = getOutputName()
+    val filename = getOutputName()
     val philosName = List("Descartes", "Seneca", "Rousseau", "Nietzsche", "Hobbes")
 
-    val logger = system.actorOf(Props(classOf[DinnerLogger], fn, duration.fromNow), "Dinner_Logger")
+    val logger = system.actorOf(Props(classOf[DinnerLogger], filename, duration.fromNow), "Dinner_Logger")
     val forks = createForks(philosName, logger)
     val philos = createPhilosophers(philosName, forks, logger)
 
+    initializePhilos(philos)
+    initializeForks(forks)
     logger ! WriteProgressBar
     philos.foreach{ _ ! Think }
     system.scheduler.scheduleOnce(duration, self, TerminateDinner(philos, forks, logger))
@@ -57,21 +59,33 @@ class ResourceHierarchyDinnerMaster(duration: FiniteDuration, fakeRandom: Boolea
     val philosQ: Int  = philoNames.length
     return for { (name, position) <- philoNames.zipWithIndex } yield createPhilosopher(
       "%s--%d".format(name, position),
-      forks(position),
+      forks(position), 
       forks((position + 1) % philosQ),
-      if ((position + 1) == philosQ) "right-handed" else "left-handed",
       logger,
       getRandomSecond(),
       getRandomSecond())
   }
 
-  private def createPhilosopher(name: String, left: ActorRef, right: ActorRef, handedness: String, logger: ActorRef, thinkingTime: FiniteDuration, eatingTime: FiniteDuration): ActorRef = {
-    return system.actorOf(Props(classOf[Philosopher], name, left, right, handedness, thinkingTime, eatingTime, logger), name)
+  private def createPhilosopher(name: String, left: ActorRef, right: ActorRef, logger: ActorRef, thinkingTime: FiniteDuration, eatingTime: FiniteDuration): ActorRef = {
+    return system.actorOf(Props(classOf[Philosopher], name, left, right, thinkingTime, eatingTime, logger), name)
+  }
+
+  private def initializePhilos(philos: List[ActorRef]): Unit = {
+    philos(0) ! InitializePhilo(false, true, false, true, philos(4), philos(1))
+    philos(1) ! InitializePhilo(true, false, false, true, philos(0), philos(2))
+    philos(2) ! InitializePhilo(true, false, false, true, philos(1), philos(3))
+    philos(3) ! InitializePhilo(true, false, false, true, philos(2), philos(4))
+    philos(4) ! InitializePhilo(true, false, true, false, philos(3), philos(0))
+  }
+
+  private def initializeForks(forks: IndexedSeq[ActorRef]): Unit = {
+    val dirty: Boolean = true
+    forks.foreach { _ ! InitializeFork(dirty) }
   }
 
   private def getOutputName(): String = {
     val option: String = if(fakeRandom) "Y" else "N"
-    return "ResourceHierarchyDinnerMaster-T%s-o%s-%s.json".format(duration.toSeconds, option, System.currentTimeMillis.toString)
+    return "ChandyMisraDinnerMaster-T%s-o%s-%s.json".format(duration.toSeconds, option, System.currentTimeMillis.toString)
   }
 
 
@@ -79,8 +93,8 @@ class ResourceHierarchyDinnerMaster(duration: FiniteDuration, fakeRandom: Boolea
     if (fakeRandom) 5.seconds else (r.nextInt(10) + 1).seconds
   }
 
-  override def postStop {
-    println("ResourceHierarchyDinnerMaster Finishing!")
+  override def postStop { 
+    println("ChandyMisraDinnerMaster Finishing!")
     system.shutdown
   }
 }
